Unified scripts’ types parsing.

+++ src/formats/map/parse/prototypes/prototype.rs
@@ -80,7 +80,7 @@ Result<prototype::Instance, errors::Error> {
         for _ in usize::MIN..overhead { inventory.push(None) }
 
         // Now this operation is safe from panic
-        inventory[index as usize] = Some(self::instance(source, provider, read_ladders_map)?);
+        inventory[index] = Some(self::instance(source, provider, read_ladders_map)?);
     }
 
     Ok(prototype::Instance {
